Fix Link of Pre-Requisites
Currently the link redirects to the GitHub markdown file. The pictures
on the site it would link to are currently broken.

This fixes the pictures to be displayed on the site and the properly
link to it like all other steps. With this change the links will be
consistently keep you on the dockerlabs.collabnix.com website.